# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Business Requirements AI Agent",
        "version": "0.1.0"
    },
    "paths": {
...
        "/projects": {
            "get": {
                "tags": [
                    "projects"
                ],
                "summary": "Get Projects List",
                "operationId": "get_projects_list_projects_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Search string",
                            "title": "Search"
                        },
                        "description": "Search string"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "description": "Offset",
                            "default": 0,
                            "title": "Offset"
                        },
                        "description": "Offset"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 50,
                            "exclusiveMinimum": 0,
                            "description": "Limit",
                            "default": 10,
                            "title": "Limit"
                        },
                        "description": "Limit"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectsGet"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "projects"
                ],
                "summary": "Create Project",
                "operationId": "create_project_projects_post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_create_project_projects_post"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectBase"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Agent service HTTP error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/requirements/{requirements_id}/export": {
            "get": {
                "tags": [
                    "requirements"
                ],
                "summary": "Export Requirements",
                "operationId": "export_requirements_requirements__requirements_id__export_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "requirements_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "The identifier of session",
                            "title": "Requirements Id"
                        },
                        "description": "The identifier of session"
                    },
                    {
                        "name": "file",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Export format: 'docx' or 'pdf'",
                            "default": "docx",
                            "title": "File"
                        },
                        "description": "Export format: 'docx' or 'pdf'"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_create_project_projects_post": {
                "properties": {
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "files": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string",
                                    "format": "binary"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Files"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "description"
                ],
                "title": "Body_create_project_projects_post"
            },
...
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Error Message"
                    },
                    "detail": {
                        "type": "string",
                        "title": "Error Detail"
                    }
                },
                "type": "object",
                "required": [
                    "message",
                    "detail"
                ],
                "title": "ErrorResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ProjectBase": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "external_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "External Id"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ProjectStatusEnum"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated At"
                    },
                    "files": {
                        "items": {
                            "$ref": "#/components/schemas/ProjectFileBase"
                        },
                        "type": "array",
                        "title": "Files",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "status",
                    "title",
                    "description",
                    "created_at",
                    "updated_at"
                ],
                "title": "ProjectBase"
            },
            "ProjectFileBase": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "file_path": {
                        "type": "string",
                        "title": "File Path"
                    },
                    "original_name": {
                        "type": "string",
                        "title": "Original Name"
                    },
                    "file_size": {
                        "type": "integer",
                        "title": "File Size"
                    },
                    "mime_type": {
                        "type": "string",
                        "title": "Mime Type"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "file_path",
                    "original_name",
                    "file_size",
                    "mime_type",
                    "created_at"
                ],
                "title": "ProjectFileBase"
            },
            "ProjectShallow": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ProjectStatusEnum"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "status",
                    "title",
                    "created_at",
                    "updated_at"
                ],
                "title": "ProjectShallow"
            },
            "ProjectStatusEnum": {
                "type": "string",
                "enum": [
                    "active",
                    "finished"
                ],
                "title": "ProjectStatusEnum"
            },
            "ProjectsGet": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/ProjectShallow"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    }
                },
                "type": "object",
                "required": [
                    "items",
                    "total"
                ],
                "title": "ProjectsGet"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}